backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-multipart
pytesseract
Pillow
pdf2image
requests